requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
opencv-python==4.8.1.78
ultralytics==8.0.200
Pillow==10.0.1
numpy==1.24.3
google-generativeai==0.3.2
python-socketio==5.9.0
eventlet==0.33.3
